Restructure to separate type-defs from package contents
