+++ libs/@local/status/README.md
@@ -19,6 +19,16 @@ A consistent structure and pattern for producing API-agnostic definitions of sta
 
 - The propagation of errors purely _within_ a service do not have to conform to the shape of the model defined here. This is a compromise to attempt to maintain idiomatic behaviour within various environments. Therefore `error-stack` can continue to be used consistently in Rust, and TS can throw errors that will be caught in a tight scope when required. However, once a path allows for that error to be propagated through an API, type systems should be used to ensure that all errors in their are defined in a publicly consumable fashion.
 
-### Attributions
+## Package Layout
+
+This package is structured into two main areas:
+
+- [`pkg`](./pkg) contains the TypeScript package that exports the `Status`, `StatusCode` types, and helper functions.
+- [`crate`](./crate) contains the Rust crate that defines the `Status` and `StatusCode` types.
+- [`type-defs`](./type-defs) contains the plain type definitions for `Status`, `StatusCode`, and associated status payloads. These types are defined in TypeScript at the moment but could easily be represented in another schema format.
+
+Note: despite the `type-defs` being in TypeScript, we define them separately to keep a better separation of concerns, and to avoid `quicktype` breaking when it encounters non-type code (e.g. `const` definitions).
+
+## Attributions
 
 - This model was heavily inspired by [Google Cloud’s error model](https://cloud.google.com/apis/design/errors)
