+++ libs/@local/status/pkg/src/main.ts
@@ -3,20 +3,5 @@
  * and RPC APIs.
  */
 
-import { StatusCode } from "./status-code";
-
-export { StatusCode } from "./status-code";
-
-/**
- * The canonical shape of a response object describing the status of a request between services.
- */
-export type Status<D extends object> = {
-  code: StatusCode;
-  /**
-   * A developer-facing description of the status.
-   *
-   * Where possible, this should provide guiding advice for debugging and/or handling the error.
-   */
-  message?: string;
-  contents: D[];
-};
+export { Status } from "../../type-defs/status";
+export { convertStatusCodeToHttpCode, StatusCode } from "./status-code";
